-added .env.example -in src/pages/api I made endpoints -finished main page  to fetch it made two route ua and uk -ua main page  add content in server/models/mainPageModels mainPageSeed.js -added collections get all, get one, create , update
