+++ .env.example
@@ -1,0 +1,10 @@
+MONGODB_URI= 'total way'
+
+# CLOUDINARY
+API_KEY= 'example'
+API_SECRET='example'
+CLOUD_NAME='example'
+
+APP_DOMAIN=http://localhost:3000  ' now it's local server'
+ENABLE_CLOUDINARY=true
+
