+++ src/app/server/lib/getQueryClient.js
@@ -4,3 +4,4 @@ import { cache } from 'react';
 const getQueryClient = cache(() => new QueryClient());
 
 export default getQueryClient;
+//TODO not yet used with component
