+++ src/app/server/lib/saveFileToUploadDir.js
@@ -10,8 +10,6 @@ const saveFileToUploadDir = async (file) => {
   const tempPath = path.join(TEMP_UPLOAD_DIR, file.filename);
   const uploadPath = path.join(UPLOAD_DIR, file.filename);
 
-  console.log('Temp Path:', tempPath);
-  console.log('Upload Path:', uploadPath);
 
   await fs.rename(tempPath, uploadPath);
