+++ src/app/server/models/SectionSchema.js
@@ -1,37 +1,44 @@
 import mongoose from 'mongoose';
+import { CollectionModel } from './CollectionsModel.js';
 
-const progressSchema = new mongoose.Schema({
-  title: { type: String, default: '' },
-  text: { type: String, default: '' },
-  item: { type: String, default: '' },
-  total: { type: String, default: '' },
-  currency: { type: String, default: '' },
-  term: { type: String, default: '' },
-  days: { type: String, default: '' },
-  period: { type: String, default: '' },
-  imageUrl: { type: String, default: '' },
-  type: { type: String, required: true },
-});
+//UNCOMMENT TO USE SEEDS FOR MAIN PAGE
 
-export const ProgressModel = mongoose.model('Progress', progressSchema);
+// const progressSchema = new mongoose.Schema({
+//   title: { type: String, default: '' },
+//   text: { type: String, default: '' },
+//   item: { type: String, default: '' },
+//   total: { type: String, default: '' },
+//   currency: { type: String, default: '' },
+//   term: { type: String, default: '' },
+//   days: { type: String, default: '' },
+//   period: { type: String, default: '' },
+//   image: { type: String },
+//   type: { type: String },
+// });
 
-const collectionSchema = new mongoose.Schema({
-  title: { type: String },
-  importance: { type: String },
-  image: { type: [String], default: [] },
-  collected: { type: Number, default: 0 },
-  target: { type: Number },
-  alt: { type: String, default: '' },
-  peopleDonate: { type: Number, default: 0 },
-  desc: { type: String },
-  long_desc: { type: String, default: '' },
-  status: { type: String, enum: ['active', 'closed'], default: 'active' },
-  createdAt: { type: Date, default: Date.now },
-  closedAt: { type: Date, default: null },
-  type: { type: String, required: true },
-});
+// export const ProgressModel = mongoose.model('Progress', progressSchema);
 
-export const CollectionModel = mongoose.model('Collection', collectionSchema);
+// const collectionSchema = new mongoose.Schema({
+//   title: { type: String },
+//   importance: { type: String },
+//   image: { type: [String], default: [] },
+//   collected: { type: Number, default: 0 },
+//   target: { type: Number },
+//   alt: { type: String, default: '' },
+//   peopleDonate: { type: Number, default: 0 },
+//   desc: { type: String },
+//   long_desc: { type: String, default: '' },
+//   status: { type: String, enum: ['active', 'closed'], default: 'active' },
+//   createdAt: { type: Date, default: Date.now },
+//   closedAt: { type: Date, default: null },
+//   type: {
+//     type: String,
+//     required: true,
+//     enum: ['collections', 'progress', 'hero', 'about', 'none'],
+//   },
+// });
+
+// export const CollectionModel = mongoose.model('Collections', collectionSchema);
 
 const sectionSchema = new mongoose.Schema({
   local: { type: String, required: true },
