+++ src/app/server/models/index.js
@@ -1,2 +1,2 @@
 export { default as SectionsSchema } from './SectionSchema.js';
-export { CollectionModel } from './SectionSchema.js';
+// export { CollectionModel } from './SectionSchema.js';
