+++ src/app/server/utils/isValidId.js
@@ -1,8 +1,7 @@
 import { isValidObjectId } from 'mongoose';
 import createHttpError from 'http-errors';
-const isValidId = (req, res, next) => {
+const isValidId = (req) => {
   const { id } = req.query;
-  console.log('Id valid', id);
   if (!isValidObjectId(id)) {
     throw createHttpError(404, 'Not found');
   }
