+++ src/pages/api/collections/[id].js
@@ -6,7 +6,7 @@ import {
 } from '@/app/server/controllers/collectionsController.js';
 import isValidId from '@/app/server/utils/isValidId.js';
 import validateBody from '@/app/server/utils/validateBody.js';
-import { CollectionModel } from '@/app/server/models/SectionSchema.js';
+import CollectionModel from '@/app/server/models/CollectionsModel.js';
 export const config = {
   api: {
     bodyParser: false,
